refactor: migrate lfs and clean up vfs afterwards
